requirements.txt
Django~=3.2
djangorestframework~=3.12.4
PyMySQL~=1.0.2